requirements.txt
asgiref==3.8.1
Django==5.0.4
python-dotenv==1.0.1
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
